Rewrite getVsEntryRegInfo
The function getVsEntryRegInfo is a little long, and mixes a lot of word
together.  There is also a problem because on GFX9 the HW shader stage
does not match the user data registers as expected.  That is, a GS
shader uses the ES user data entries.

This PR will rewrite that code and fix up this problem.